Feature #1: Implement Enter Session feature
Adding the notion of "Notes" to the project. When an user selects a
session, a request is done to the server to load all the notes from
that session.

The user name is trying to be read from the NSUserDefaults. If it's
null, all the notes from the sessions are loaded, if it isn't null,
it only loads the notes from that user.