Improve add note screen and dark mode issue
Improved add note screen to be a little bit more similar to what
will be in future. It depends on the topic passed by the previous
screen for the final version, so for now, I left the view picker.

Set the colors manually so dark mode doesn't make things invisible.